+++ digiwf-engine/digiwf-engine-service/src/main/java/io/muenchendigital/digiwf/legacy/document/domain/DocumentService.java
@@ -11,12 +11,16 @@ import io.muenchendigital.digiwf.process.definition.domain.service.ServiceDefini
 import io.muenchendigital.digiwf.shared.exception.ObjectNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import lombok.val;
 import org.camunda.bpm.engine.RuntimeService;
+import org.camunda.bpm.engine.TaskService;
 import org.camunda.bpm.engine.repository.ProcessDefinition;
 import org.camunda.bpm.engine.task.Task;
 import org.springframework.stereotype.Service;
 
 import java.util.Map;
+
+import static io.muenchendigital.digiwf.task.TaskVariables.TASK_ASSIGNEE;
 
 @Slf4j
 @Service
@@ -26,6 +30,7 @@ public class DocumentService {
     private final CosysClient cosysClient;
     private final ServiceDefinitionService processDefinitionService;
     private final HumanTaskService taskService;
+    private final TaskService camundaTaskService;
     private final ProcessConfigService processConfigService;
 
     private final RuntimeService runtimeService;
@@ -38,10 +43,11 @@ public class DocumentService {
         return this.cosysClient.generateDocument(documentGuid, this.runtimeService.getVariables(executionId));
     }
 
-    public byte[] getStatusDokumentForTask(final String taskId, final String userLogin) {
+    public byte[] getStatusDokumentForTask(final String taskId, final String userId) {
         final Task task = this.taskService.getTask(taskId);
-        if (!userLogin.equals(task.getAssignee())) {
-            throw new IllegalArgumentException(String.format("The task with the id %s is not available.", taskId));
+        val assignedUserId = TASK_ASSIGNEE.from(camundaTaskService, taskId).getLocalOptional().orElseGet(task::getAssignee);
+        if (!userId.equals(assignedUserId)) {
+            throw new ObjectNotFoundException(String.format("The task with the id %s is not available.", taskId));
         }
 
         final ProcessDefinition processDefinition = this.processDefinitionService.getServiceDefinition(task.getProcessDefinitionId());