+++ digiwf-integrations/digiwf-email-integration/digiwf-email-integration-core/src/main/java/io/muenchendigital/digiwf/email/integration/model/FileAttachment.java
@@ -1,0 +1,18 @@
+package io.muenchendigital.digiwf.email.integration.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.RequiredArgsConstructor;
+
+import javax.mail.util.ByteArrayDataSource;
+
+@Data
+@RequiredArgsConstructor
+@AllArgsConstructor
+public class FileAttachment {
+
+    private String fileName;
+
+    private ByteArrayDataSource file;
+
+}