+++ digiwf-integrations/digiwf-email-integration/digiwf-email-integration-example/src/main/java/io/muenchendigital/digiwf/email/integration/DigiwfEmailExampleApplication.java
@@ -1,0 +1,13 @@
+package io.muenchendigital.digiwf.email.integration;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DigiwfEmailExampleApplication {
+
+    public static void main(final String[] args) {
+        SpringApplication.run(DigiwfEmailExampleApplication.class, args);
+    }
+
+}