+++ digiwf-integrations/digiwf-email-integration/digiwf-email-integration-example/src/main/java/io/muenchendigital/digiwf/email/integration/api/controller/ExampleController.java
@@ -1,32 +1,22 @@
 package io.muenchendigital.digiwf.email.integration.api.controller;
 
-import io.muenchendigital.digiwf.email.integration.domain.model.Mail;
-import io.muenchendigital.digiwf.email.integration.domain.service.MailingService;
-import io.muenchendigital.digiwf.spring.cloudstream.utils.api.streaming.service.PayloadSenderService;
+import io.muenchendigital.digiwf.email.integration.model.Mail;
+import io.muenchendigital.digiwf.message.common.MessageConstants;
+import io.muenchendigital.digiwf.message.core.api.MessageApi;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
-import javax.validation.ConstraintViolationException;
+import java.util.Map;
 
 @RestController
 @RequiredArgsConstructor
 @Slf4j
 public class ExampleController {
 
-    private final MailingService mailingService;
-    private final PayloadSenderService genericPayloadSender;
-
-    @PostMapping(value = "/testSendMail")
-    public void testSendMail(@RequestBody final Mail mail) {
-        try {
-            this.mailingService.sendMail(mail);
-        } catch (final ConstraintViolationException e) {
-            log.error(e.toString());
-        }
-    }
+    private final MessageApi messageApi;
 
     /**
      * Note: for this to work, you have to configure both
@@ -36,7 +26,11 @@ public class ExampleController {
      */
     @PostMapping(value = "/testEventBus")
     public void testEventBus(@RequestBody final Mail mail) {
-        this.genericPayloadSender.sendPayload(mail, "sendMailFromEventBus");
+        this.messageApi.sendMessage(mail, Map.of(
+                MessageConstants.TYPE, "sendMailFromEventBus",
+                MessageConstants.DIGIWF_PROCESS_INSTANCE_ID, "processInstanceId",
+                MessageConstants.DIGIWF_MESSAGE_NAME, "testEmailIntegration"
+        ), "dwf-email-local-01");
     }
 
 }