+++ digiwf-integrations/digiwf-s3-integration/digiwf-s3-integration-core/src/main/java/io/muenchendigital/digiwf/s3/integration/application/CreatePresignedUrlsUseCase.java
@@ -1,0 +1,35 @@
+package io.muenchendigital.digiwf.s3.integration.application;
+
+import io.minio.http.Method;
+import io.muenchendigital.digiwf.s3.integration.adapter.in.streaming.CreatePresignedUrlEvent;
+import io.muenchendigital.digiwf.s3.integration.application.port.in.CreatePresignedUrlsInPort;
+import io.muenchendigital.digiwf.s3.integration.application.port.in.FileSystemAccessException;
+import io.muenchendigital.digiwf.s3.integration.domain.model.PresignedUrl;
+import lombok.RequiredArgsConstructor;
+import org.springframework.lang.NonNull;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.validation.annotation.Validated;
+
+import javax.validation.Valid;
+import java.util.List;
+
+@RequiredArgsConstructor
+@Validated
+@Transactional
+public class CreatePresignedUrlsUseCase implements CreatePresignedUrlsInPort {
+
+  private final FileOperationsUseCase fileHandlingService;
+  private final int presignedUrlExpiresInMinutes;
+
+  @Override
+  @NonNull
+  public List<PresignedUrl> createPresignedUrls(@Valid CreatePresignedUrlEvent event) throws FileSystemAccessException {
+    // No end of life is set for files to be saved
+    return this.fileHandlingService.getPresignedUrls(
+        List.of(event.getPath().split(";")),
+        Method.valueOf(event.getAction()),
+        this.presignedUrlExpiresInMinutes // 7 days is max expiration
+    );
+  }
+
+}