+++ digiwf-integrations/digiwf-s3-integration/digiwf-s3-integration-service/src/main/java/io/muenchendigital/digiwf/s3/integration/S3IntegrationApplication.java
@@ -18,10 +18,8 @@ import org.springframework.scheduling.annotation.EnableScheduling;
 @EnableScheduling
 public class S3IntegrationApplication {
 
-
-    public static void main(String[] args) {
-        SpringApplication.run(S3IntegrationApplication.class, args);
-    }
-
+  public static void main(String[] args) {
+    SpringApplication.run(S3IntegrationApplication.class, args);
+  }
 
 }