+++ digiwf-libs/digiwf-message/digiwf-message-core/src/main/java/io/muenchendigital/digiwf/message/process/api/error/BpmnError.java
@@ -1,0 +1,16 @@
+package io.muenchendigital.digiwf.message.process.api.error;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+/**
+ * Exception to be thrown when an integration fails with a bpmn error that may be handled by the caller process.
+ */
+@AllArgsConstructor
+@Getter
+public class BpmnError extends RuntimeException {
+
+    private final String errorCode;
+    private final String errorMessage;
+
+}